requirements.txt
Django==1.6
Fabric==1.8.0
Markdown==2.3.1
South==0.8.2
amqp==1.0.13
anyjson==0.3.3
billiard==2.7.3.34
boto==2.17.0
celery==3.0.24
django-devserver==0.6.2
django-filter==0.7
django-registration==1.0
django-storages==1.1.8
djangorestframework==2.3.8
ecdsa==0.9
kombu==2.5.15
paramiko==1.12.0
psycopg2==2.5.1
pycrypto==2.6.1
python-dateutil==2.1
six==1.4.1
wsgiref==0.1.2
